+++ clients/w3wprediction/src/restapi/models/model_get_otc_blocktrade_detail_response_order_data.go
@@ -26,7 +26,7 @@ type GetOtcBlocktradeDetailResponseOrderData struct {
 	Taker                *string  `json:"taker,omitempty"`
 	MakerAmount          *string  `json:"makerAmount,omitempty"`
 	TakerAmount          *string  `json:"takerAmount,omitempty"`
-	Price                *float32 `json:"price,omitempty"`
+	Price                *float64 `json:"price,omitempty"`
 	OrderType            *string  `json:"orderType,omitempty"`
 	TimeInForce          *string  `json:"timeInForce,omitempty"`
 	Expiration           *int64   `json:"expiration,omitempty"`
@@ -316,9 +316,9 @@ func (o *GetOtcBlocktradeDetailResponseOrderData) SetTakerAmount(v string) {
 }
 
 // GetPrice returns the Price field value if set, zero value otherwise.
-func (o *GetOtcBlocktradeDetailResponseOrderData) GetPrice() float32 {
+func (o *GetOtcBlocktradeDetailResponseOrderData) GetPrice() float64 {
 	if o == nil || common.IsNil(o.Price) {
-		var ret float32
+		var ret float64
 		return ret
 	}
 	return *o.Price
@@ -326,7 +326,7 @@ func (o *GetOtcBlocktradeDetailResponseOrderData) GetPrice() float32 {
 
 // GetPriceOk returns a tuple with the Price field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *GetOtcBlocktradeDetailResponseOrderData) GetPriceOk() (*float32, bool) {
+func (o *GetOtcBlocktradeDetailResponseOrderData) GetPriceOk() (*float64, bool) {
 	if o == nil || common.IsNil(o.Price) {
 		return nil, false
 	}
@@ -342,8 +342,8 @@ func (o *GetOtcBlocktradeDetailResponseOrderData) HasPrice() bool {
 	return false
 }
 
-// SetPrice gets a reference to the given float32 and assigns it to the Price field.
-func (o *GetOtcBlocktradeDetailResponseOrderData) SetPrice(v float32) {
+// SetPrice gets a reference to the given float64 and assigns it to the Price field.
+func (o *GetOtcBlocktradeDetailResponseOrderData) SetPrice(v float64) {
 	o.Price = &v
 }